enable multiple train mode SSDU original and FANSS

+++ generate_output.py
@@ -12,8 +12,10 @@ def display_output(ssdu_model, test_dataset, image_save_folder):
     def myprint(s):
         with open(os.path.join(image_save_folder, 'modelsummary.txt'),'a') as f:
             print(s, file=f)
-    
-    os.remove(os.path.join(image_save_folder, 'modelsummary.txt'))
+
+    if os.path.exists(os.path.join(image_save_folder, 'modelsummary.txt')):
+        os.remove(os.path.join(image_save_folder, 'modelsummary.txt'))
+
     ssdu_model.summary(print_fn=myprint)
     for element in test_dataset.take(1):
         nw_input, sens_maps, trn_mask, loss_mask = element[0]
@@ -27,7 +29,7 @@ def display_output(ssdu_model, test_dataset, image_save_folder):
         fig1.savefig(os.path.join(image_save_folder, 'nw_input.png'))
 
         #display network output image    
-        output = ssdu_model(element[0], training = False)
+        output = ssdu_model.predict(element[0])
         output = output[0]
         fig2 = plt.figure()
         output = tf_utils.tf_real2complex(output).numpy()[0, 0,...]
@@ -69,11 +71,11 @@ def SSDU_inference(ssdu_model, inference_dataset, patient_folder, num_slices=350
     for element in inference_dataset.take(num_slices):
         print('Performing Inference for Slice ', str(iSlice), '. Patient folder:  ', patient_folder)
         #get network output image    
-        output = ssdu_model(element[0], training = False)
+        output = ssdu_model.predict(element[0])
         output = output[0]
         output = tf_utils.tf_real2complex(output).numpy()[0, ...] #shape (1, nMaps, nRow, nCol)
         output = np.squeeze(output) #shape (nMaps, nRow, nCol)
-        output = np.transpose(output, (1, 2, 0)) #shape (nMaps, nRow, nCol)
+        output = np.transpose(output, (1, 2, 0)) #shape (nRow, nCol, nMaps)
 
         save_dict = {"SC_12_SSDU": output}
         scipy.io.savemat(os.path.join(DL_inference_folder,'Slice_' + str(iSlice) + '.mat'), save_dict)
